feat: added stars badge for symfony insight
