+++ lib/text-formatters.js
@@ -7,7 +7,7 @@
 const moment = require('moment')
 moment().format()
 
-function starRating(rating) {
+function starRating(rating, max = 5) {
   const flooredRating = Math.floor(rating)
   let stars = ''
   while (stars.length < flooredRating) {
@@ -23,7 +23,8 @@ function starRating(rating) {
   } else if (decimal >= 0.125) {
     stars += '¼'
   }
-  while (stars.length < 5) {
+
+  while (stars.length < max) {
     stars += '☆'
   }
   return stars
