+++ lib/text-formatters.spec.js
@@ -23,6 +23,7 @@ describe('Text formatters', function() {
     given(2.566).expect('★★½☆☆')
     given(2.2).expect('★★¼☆☆')
     given(3).expect('★★★☆☆')
+    given(2, 4).expect('★★☆☆')
   })
 
   test(currencyFromCode, () => {
