+++ services/symfony/sensiolabs-redirect.service.js
@@ -1,0 +1,17 @@
+'use strict'
+
+const { redirector } = require('..')
+
+module.exports = [
+  // The SymfonyInsight service was previously branded as SensioLabs, and
+  // accordingly the badge path used to be /sensiolabs/i/projectUuid'.
+  redirector({
+    category: 'analysis',
+    route: {
+      base: 'sensiolabs/i',
+      pattern: ':projectUuid',
+    },
+    target: ({ projectUuid }) => `/symfony/i/grade/${projectUuid}`,
+    dateAdded: new Date('2019-02-08'),
+  }),
+]
