+++ services/symfony/sensiolabs-redirect.tester.js
@@ -1,0 +1,18 @@
+'use strict'
+
+const { ServiceTester } = require('../tester')
+
+const t = (module.exports = new ServiceTester({
+  id: 'sensiolabs',
+  title: 'SensioLabs',
+}))
+
+t.create('sensiolabs insight')
+  .get('/i/45afb680-d4e6-4e66-93ea-bcfa79eb8a87.svg', {
+    followRedirect: false,
+  })
+  .expectStatus(301)
+  .expectHeader(
+    'Location',
+    '/symfony/i/grade/45afb680-d4e6-4e66-93ea-bcfa79eb8a87.svg'
+  )
